git-svn-id: http://joomlacode.org/svn/joomla/development/branches/jean_marie@18227 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ administrator/components/com_users/models/user.php
@@ -8,7 +8,7 @@
 // No direct access.
 defined('_JEXEC') or die;
 
-jimport('joomla.application.component.modelform');
+jimport('joomla.application.component.modeladmin');
 
 /**
  * User model.
@@ -17,33 +17,8 @@ jimport('joomla.application.component.modelform');
  * @subpackage	com_users
  * @since		1.6
  */
-class UsersModelUser extends JModelForm
+class UsersModelUser extends JModelAdmin
 {
-	/**
-	 * Method to auto-populate the model state.
-	 */
-	protected function _populateState()
-	{
-		$app = JFactory::getApplication('administrator');
-
-		// Load the User state.
-		if (!($pk = (int) $app->getUserState('com_users.edit.user.id'))) {
-			$pk = (int) JRequest::getInt('id');
-		}
-		$this->setState('user.id', $pk);
-
-		// Load the parameters.
-		$params	= JComponentHelper::getParams('com_users');
-		$this->setState('params', $params);
-	}
-
-	/**
-	 * Prepare and sanitise the table prior to saving.
-	 */
-	protected function _prepareTable(&$table)
-	{
-	}
-
 	/**
 	 * Returns a reference to the a Table object, always creating it.
 	 *
@@ -51,6 +26,7 @@ class UsersModelUser extends JModelForm
 	 * @param	string	A prefix for the table class name. Optional.
 	 * @param	array	Configuration array for model. Optional.
 	 * @return	JTable	A database object
+	 * @since	1.6
 	*/
 	public function getTable($type = 'User', $prefix = 'JTable', $config = array())
 	{
@@ -62,92 +38,87 @@ class UsersModelUser extends JModelForm
 	 * Method to get a single record.
 	 *
 	 * @param	integer	The id of the primary key.
-	 *
 	 * @return	mixed	Object on success, false on failure.
+	 * @since	1.6
 	 */
-	public function &getItem($pk = null)
+	public function getItem($pk = null)
 	{
-		// Initialise variables.
-		$pk = (!empty($pk)) ? $pk : (int)$this->getState('user.id');
-		$false	= false;
-
-		// Get a row instance.
-		$table = &$this->getTable();
-
-		// Attempt to load the row.
-		$return = $table->load($pk);
-
-		// Check for a table object error.
-		if ($return === false && $table->getError()) {
-			$this->setError($table->getError());
-			return $false;
-		}
-
-		// Prime required properties.
-		if (empty($table->id))
-		{
-			// Prepare data for a new record.
-		}
-
-		// Convert to the JObject before adding other data.
-		$value = JArrayHelper::toObject($table->getProperties(1), 'JObject');
-		$value->profile = new JObject;
+		$result = parent::getItem($pk);
 
 		// Get the dispatcher and load the users plugins.
-		$dispatcher	= &JDispatcher::getInstance();
+		$dispatcher	= JDispatcher::getInstance();
 		JPluginHelper::importPlugin('user');
 
 		// Trigger the data preparation event.
-		$results = $dispatcher->trigger('onPrepareUserProfileData', array($table->id, &$value));
+		$results = $dispatcher->trigger('onContentPrepareData', array('com_users.user', $result));
 
-		// Convert the params field to an array.
-		$registry = new JRegistry;
-		$registry->loadJSON($value->params);
-		$value->params = $registry->toArray();
-
-		return $value;
+		return $result;
 	}
 
 	/**
 	 * Method to get the record form.
 	 *
-	 * @return	mixed	JForm object on success, false on failure.
+	 * @param	array	$data		An optional array of data for the form to interogate.
+	 * @param	boolean	$loadData	True if the form is to load its own data (default case), false if not.
+	 * @return	JForm	A JForm object on success, false on failure
+	 * @since	1.6
 	 */
-	public function getForm()
+	public function getForm($data = array(), $loadData = true)
 	{
 		// Initialise variables.
 		$app = JFactory::getApplication();
 
 		// Get the form.
-		try {
-			$form = parent::getForm('com_users.user', 'user', array('control' => 'jform'));
-		} catch (Exception $e) {
-			$this->setError($e->getMessage());
+		$form = $this->loadForm('com_users.user', 'user', array('control' => 'jform', 'load_data' => $loadData));
+		if (empty($form)) {
 			return false;
-		}
-
-		// Get the dispatcher and load the users plugins.
-		$dispatcher	= &JDispatcher::getInstance();
-		JPluginHelper::importPlugin('user');
-
-		// Trigger the form preparation event.
-		$results = $dispatcher->trigger('onPrepareUserProfileForm', array($this->getState('user.id'), &$form));
-
-		// Check for errors encountered while preparing the form.
-		if (count($results) && in_array(false, $results, true)) {
-			$this->setError($dispatcher->getError());
-			return false;
-		}
-
-		// Check the session for previously entered form data.
-		$data = $app->getUserState('com_users.edit.user.data', array());
-
-		// Bind the form data if present.
-		if (!empty($data)) {
-			$form->bind($data);
 		}
 
 		return $form;
+	}
+
+	/**
+	 * Method to get the data that should be injected in the form.
+	 *
+	 * @return	mixed	The data for the form.
+	 * @since	1.6
+	 */
+	protected function loadFormData()
+	{
+		// Check the session for previously entered form data.
+		$data = JFactory::getApplication()->getUserState('com_users.edit.user.data', array());
+
+		if (empty($data)) {
+			$data = $this->getItem();
+		}
+
+		// TODO: Maybe this can go into the parent model somehow?
+		// Get the dispatcher and load the users plugins.
+		$dispatcher	= JDispatcher::getInstance();
+		JPluginHelper::importPlugin('user');
+
+		// Trigger the data preparation event.
+		$results = $dispatcher->trigger('onContentPrepareData', array('com_users.profile', $data));
+
+		// Check for errors encountered while preparing the data.
+		if (count($results) && in_array(false, $results, true)) {
+			$this->setError($dispatcher->getError());
+		}
+
+		return $data;
+	}
+
+	/**
+	 * Override preprocessForm to load the user plugin group instead of content.
+	 *
+	 * @param	object	A form object.
+	 * @param	mixed	The data expected for the form.
+	 * @throws	Exception if there is an error in the form event.
+	 * @since	1.6
+	 */
+	protected function preprocessForm(JForm $form, $data)
+	{
+		parent::preprocessForm($form, $data, 'user');
 	}
 
 	/**
@@ -155,6 +126,7 @@ class UsersModelUser extends JModelForm
 	 *
 	 * @param	array	The form data.
 	 * @return	boolean	True on success.
+	 * @since	1.6
 	 */
 	public function save($data)
 	{
@@ -168,40 +140,34 @@ class UsersModelUser extends JModelForm
 		JPluginHelper::importPlugin('user');
 
 		// Load the row if saving an existing record.
-		if ($pk > 0)
-		{
+		if ($pk > 0) {
 			$table->load($pk);
 			$isNew = false;
 		}
 
 		// The password field is a special case.
-		if (!empty($data['password']))
-		{
+		if (!empty($data['password'])) {
 			// Generate a password hash.
 			jimport('joomla.user.helper');
 			$salt  = JUserHelper::genRandomPassword(32);
 			$crypt = JUserHelper::getCryptedPassword($data['password'], $salt);
 			$data['password'] = $crypt.':'.$salt;
-		}
-		else
-		{
+		} else {
 			// Do nothing to the password field.
 			unset($data['password']);
 		}
 
 		// Bind the data.
-		if (!$table->bind($data))
-		{
+		if (!$table->bind($data)) {
 			$this->setError($table->getError());
 			return false;
 		}
 
 		// Prepare the row for saving.
-		$this->_prepareTable($table);
+		$this->prepareTable($table);
 
 		// Check the data.
-		if (!$table->check())
-		{
+		if (!$table->check()) {
 			$this->setError($table->getError());
 			return false;
 		}
@@ -212,28 +178,34 @@ class UsersModelUser extends JModelForm
 		// Merge the table back into the raw data for plugin processing.
 		$data = array_merge($data, $table->getProperties(true));
 
-		// Trigger the onBeforeStoreUser event.
-		$result = $dispatcher->trigger('onBeforeStoreUser', array($old->getProperties(true), $isNew, $data));
+		// Trigger the onUserBeforeSave event.
+		$result = $dispatcher->trigger('onUserBeforeSave', array($old->getProperties(true), $isNew, $data));
 		if (in_array(false, $result, true)) {
 			$this->setError($table->getError());
 			return false;
 		}
 
 		// Store the data.
-		if (!$table->store())
-		{
+		if (!$table->store()) {
 			$this->setError($table->getError());
 			return false;
 		}
 
-		$user = &JFactory::getUser();
+		// Insert the ID into the data if this is a new user.
+		// Otherwise the AfterSave plugins won't work.
+		if ($isNew) {
+			$data['id'] = $table->id;
+		}
+
+		$user = JFactory::getUser();
 		if ($user->id == $table->id) {
 			$registry = new JRegistry;
 			$registry->loadJSON($table->params);
 			$user->setParameters($registry);
 		}
+
 		// Trigger the onAftereStoreUser event
-		$dispatcher->trigger('onAfterStoreUser', array($data, $isNew, true, null));
+		$dispatcher->trigger('onUserAfterSave', array($data, $isNew, true, null));
 
 		$this->setState('user.id', $table->id);
 
@@ -244,8 +216,8 @@ class UsersModelUser extends JModelForm
 	 * Method to delete rows.
 	 *
 	 * @param	array	An array of item ids.
-	 *
 	 * @return	boolean	Returns true on success, false on failure.
+	 * @since	1.6
 	 */
 	public function delete(&$pks)
 	{
@@ -254,46 +226,41 @@ class UsersModelUser extends JModelForm
 		$table	= $this->getTable();
 		$pks	= (array) $pks;
 
-		// Trigger the onBeforeStoreUser event.
+		// Trigger the onUserBeforeSave event.
 		JPluginHelper::importPlugin('user');
-		$dispatcher = &JDispatcher::getInstance();
+		$dispatcher = JDispatcher::getInstance();
+
+		if (in_array($user->id, $pks)) {
+			$this->setError(JText::_('COM_USERS_USERS_ERROR_CANNOT_DELETE_SELF'));
+			return false;
+		}
 
 		// Iterate the items to delete each one.
-		foreach ($pks as $i => $pk)
-		{
-			if ($table->load($pk))
-			{
+		foreach ($pks as $i => $pk) {
+			if ($table->load($pk)) {
 				// Access checks.
 				$allow = $user->authorise('core.edit.state', 'com_users');
 
-				if ($allow)
-				{
+				if ($allow) {
 					// Get user data for the user to delete.
-					$user = & JFactory::getUser($pk);
+					$user = JFactory::getUser($pk);
 
-					// Fire the onBeforeDeleteUser event.
-					$dispatcher->trigger('onBeforeDeleteUser', array($table->getProperties()));
+					// Fire the onUserBeforeDelete event.
+					$dispatcher->trigger('onUserBeforeDelete', array($table->getProperties()));
 
-					if (!$table->delete($pk))
-					{
+					if (!$table->delete($pk)) {
 						$this->setError($table->getError());
 						return false;
+					} else {
+						// Trigger the onUserAfterDelete event.
+						$dispatcher->trigger('onUserAfterDelete', array($user->getProperties(), true, $this->getError()));
 					}
-					else
-					{
-						// Trigger the onAfterDeleteUser event.
-						$dispatcher->trigger('onAfterDeleteUser', array($user->getProperties(), true, $this->getError()));
-					}
-				}
-				else
-				{
+				} else {
 					// Prune items that you can't change.
 					unset($pks[$i]);
-					JError::raiseWarning(403, JText::_('JError_Core_Delete_not_permitted'));
+					JError::raiseWarning(403, JText::_('JERROR_CORE_DELETE_NOT_PERMITTED'));
 				}
-			}
-			else
-			{
+			} else {
 				$this->setError($table->getError());
 				return false;
 			}
@@ -309,6 +276,7 @@ class UsersModelUser extends JModelForm
 	 * @param	int		The value of the published state
 	 *
 	 * @return	boolean	True on success.
+	 * @since	1.6
 	 */
 	function block(&$pks, $value = 1)
 	{
@@ -322,16 +290,13 @@ class UsersModelUser extends JModelForm
 		JPluginHelper::importPlugin('user');
 
 		// Access checks.
-		foreach ($pks as $i => $pk)
-		{
-			if ($value == 1 && $pk == $user->get('id'))
-			{
+		foreach ($pks as $i => $pk) {
+			if ($value == 1 && $pk == $user->get('id')) {
 				// Cannot block yourself.
 				unset($pks[$i]);
-				JError::raiseWarning(403, JText::_('Users_Error_Cannot_block_self'));
-			}
-			else if ($table->load($pk))
-			{
+				JError::raiseWarning(403, JText::_('COM_USERS_USERS_ERROR_CANNOT_DELETE_SELF'));
+
+			} else if ($table->load($pk)) {
 				$old	= $table->getProperties();
 				$allow	= $user->authorise('core.edit.state', 'com_users');
 
@@ -340,39 +305,34 @@ class UsersModelUser extends JModelForm
 					'clientid' => array(0, 1)
 				);
 
-				if ($allow)
-				{
+				if ($allow) {
 					$table->block = (int) $value;
 
-					if (!$table->check())
-					{
+					if (!$table->check()) {
 						$this->setError($table->getError());
 						return false;
 					}
 
-					// Trigger the onBeforeStoreUser event.
-					$dispatcher->trigger('onBeforeStoreUser', array($old, false));
+					// Trigger the onUserBeforeSave event.
+					$dispatcher->trigger('onUserBeforeSave', array($old, false));
 
 					// Store the table.
-					if (!$table->store())
-					{
+					if (!$table->store()) {
 						$this->setError($table->getError());
 						return false;
 					}
 
 					// Trigger the onAftereStoreUser event
-					$dispatcher->trigger('onAfterStoreUser', array($table->getProperties(), false, true, null));
+					$dispatcher->trigger('onUserAfterSave', array($table->getProperties(), false, true, null));
 
 					// Log the user out.
 					if ($value) {
 						$app->logout($table->id, $options);
 					}
-				}
-				else
-				{
+				} else {
 					// Prune items that you can't change.
 					unset($pks[$i]);
-					JError::raiseWarning(403, JText::_('JError_Core_Edit_State_not_permitted'));
+					JError::raiseWarning(403, JText::_('JERROR_CORE_EDIT_STATE_NOT_PERMITTED'));
 				}
 			}
 		}
@@ -386,6 +346,7 @@ class UsersModelUser extends JModelForm
 	 * @param	array	The ids of the items to activate.
 	 *
 	 * @return	boolean	True on success.
+	 * @since	1.6
 	 */
 	function activate(&$pks)
 	{
@@ -396,47 +357,38 @@ class UsersModelUser extends JModelForm
 		$pks		= (array) $pks;
 
 		// Access checks.
-		foreach ($pks as $i => $pk)
-		{
-			if ($table->load($pk))
-			{
+		foreach ($pks as $i => $pk) {
+			if ($table->load($pk)) {
 				$old	= $table->getProperties();
 				$allow	= $user->authorise('core.edit.state', 'com_users');
 
-				if (empty($table->activation))
-				{
+				if (empty($table->activation)) {
 					// Ignore activated accounts.
 					unset($pks[$i]);
-				}
-				else if ($allow)
-				{
+				} else if ($allow) {
 					$table->block		= 0;
 					$table->activation	= '';
 
-					if (!$table->check())
-					{
+					if (!$table->check()) {
 						$this->setError($table->getError());
 						return false;
 					}
 
-					// Trigger the onBeforeStoreUser event.
-					$dispatcher->trigger('onBeforeStoreUser', array($old, false));
+					// Trigger the onUserBeforeSave event.
+					$dispatcher->trigger('onUserBeforeSave', array($old, false));
 
 					// Store the table.
-					if (!$table->store())
-					{
+					if (!$table->store()) {
 						$this->setError($table->getError());
 						return false;
 					}
 
 					// Fire the onAftereStoreUser event
-					$dispatcher->trigger('onAfterStoreUser', array($table->getProperties(), false, true, null));
-				}
-				else
-				{
+					$dispatcher->trigger('onUserAfterSave', array($table->getProperties(), false, true, null));
+				} else {
 					// Prune items that you can't change.
 					unset($pks[$i]);
-					JError::raiseWarning(403, JText::_('JError_Core_Edit_State_not_permitted'));
+					JError::raiseWarning(403, JText::_('JERROR_CORE_EDIT_STATE_NOT_PERMITTED'));
 				}
 			}
 		}
@@ -449,23 +401,21 @@ class UsersModelUser extends JModelForm
 	 *
 	 * @param	array	An array of variable for the batch operation
 	 * @param	array	An array of IDs on which to operate
+	 * @since	1.6
 	 */
 	public function batch($config, $user_ids)
 	{
 		// Ensure there are selected users to operate on.
-		if (empty($user_ids))
-		{
-			$this->setError(JText::_('USERS_USERS_NOT_SELECTED'));
+		if (empty($user_ids)) {
+			$this->setError(JText::_('COM_USERS_USERS_NO_ITEM_SELECTED'));
 			return false;
-		}
-		// Only run operations if a config array is present.
-		else if (!empty($config))
-		{
+		} else if (!empty($config)) {
+			// Only run operations if a config array is present.
 			// Ensure there is a valid group.
 			$group_id = JArrayHelper::getValue($config, 'group_id', 0, 'int');
-			if ($group_id < 1)
-			{
-				$this->setError(JText::_('USERS_INVALID_GROUP'));
+
+			if ($group_id < 1) {
+				$this->setError(JText::_('COM_USERS_ERROR_INVALID_GROUP'));
 				return false;
 			}
 
@@ -474,8 +424,7 @@ class UsersModelUser extends JModelForm
 			$oldAclMode = $acl->setCheckMode(1);
 
 			$groupLogic	= JArrayHelper::getValue($config, 'group_logic');
-			switch ($groupLogic)
-			{
+			switch ($groupLogic) {
 				case 'set':
 					$doDelete		= 2;
 					$doAssign		= true;
@@ -494,22 +443,18 @@ class UsersModelUser extends JModelForm
 			}
 
 			// Remove the users from the group(s) if requested.
-			if ($doDelete)
-			{
+			if ($doDelete) {
 				// Purge operation, remove the users from all groups.
-				if ($doDelete === 2)
-				{
+				if ($doDelete === 2) {
 					$this->_db->setQuery(
-						'DELETE FROM `#__core_acl_groups_aro_map`' .
-						' WHERE `aro_id` IN ('.implode(',', $user_ids).')'
+						'DELETE FROM `#__user_usergroup_map`' .
+						' WHERE `user_id` IN ('.implode(',', $user_ids).')'
 					);
-				}
-				// Remove the users from the group.
-				else
-				{
+				} else {
+					// Remove the users from the group.
 					$this->_db->setQuery(
-						'DELETE FROM `#__core_acl_groups_aro_map`' .
-						' WHERE `aro_id` IN ('.implode(',', $user_ids).')' .
+						'DELETE FROM `#__user_usergroup_map`' .
+						' WHERE `user_id` IN ('.implode(',', $user_ids).')' .
 						' AND `group_id` = '.$group_id
 					);
 				}
@@ -522,17 +467,15 @@ class UsersModelUser extends JModelForm
 			}
 
 			// Assign the users to the group if requested.
-			if ($doAssign)
-			{
+			if ($doAssign) {
 				// Build the tuples array for the assignment query.
 				$tuples = array();
-				foreach ($user_ids as $id)
-				{
+				foreach ($user_ids as $id) {
 					$tuples[] = '('.$id.','.$group_id.')';
 				}
 
 				$this->_db->setQuery(
-					'INSERT IGNORE INTO `#__core_acl_groups_aro_map` (`aro_id`, `group_id`)' .
+					'INSERT IGNORE INTO `#__user_usergroup_map` (`user_id`, `group_id`)' .
 					' VALUES '.implode(',', $tuples)
 				);
 
@@ -554,6 +497,7 @@ class UsersModelUser extends JModelForm
 	 * Gets the available groups.
 	 *
 	 * @return	array
+	 * @since	1.6
 	 */
 	public function getGroups()
 	{
@@ -565,14 +509,23 @@ class UsersModelUser extends JModelForm
 	 * Gets the groups this object is assigned to
 	 *
 	 * @return	array
+	 * @since	1.6
 	 */
 	public function getAssignedGroups($userId = null)
 	{
 		// Initialise variables.
 		$userId = (!empty($userId)) ? $userId : (int)$this->getState('user.id');
 
-		jimport('joomla.user.helper');
-		$result = JUserHelper::getUserGroups($userId);
+		if (empty($userId)) {
+			$result = array();
+			$config = JComponentHelper::getParams('com_users');
+			if ($groupId = $config->get('new_usertype')) {
+				$result[$groupId] = null;
+			}
+		} else {
+			jimport('joomla.user.helper');
+			$result = JUserHelper::getUserGroups($userId);
+		}
 
 		return $result;
 	}
